+++ src/router/Router.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
 import AccountDeletion from '../components/MyPage/AccountDeletion';
 import ChangePassword from '../components/MyPage/ChangePassword';
-import MyInfo from '../components/MyPage/MyInfo';
+import MyInfo from '../components/MyPage/MyProfile';
 import CreatePostPage from '../pages/CreatePostPage';
 import CreateProfilePage from '../pages/CreateProfilePage';
 import PostDetailPage from '../pages/DetailPage';
@@ -12,6 +12,7 @@ import MyPage from '../pages/MyPage';
 import ResetPasswordPage from '../pages/ResetPasswordPage';
 import MyMeetings from '../components/MyPage/MyMeetings';
 import ViewParticipantPage from '../pages/ViewParticipantPage';
+import UserProfilePage from '../pages/UserProfilePage';
 
 const Router = () => {
   return (
@@ -19,17 +20,24 @@ const Router = () => {
       <Route path="/" element={<ListPage />} />
       <Route path="/post/:id" element={<PostDetailPage />} />
       <Route path="/create" element={<CreatePostPage />} />
+      {/* 로그인 & 회원가입 */}
       <Route path="/login" element={<LoginPage />} />
       <Route path="/join" element={<JoinPage />} />
       <Route path="/reset-password" element={<ResetPasswordPage />} />
       <Route path="/create-profile" element={<CreateProfilePage />} />
+      {/* 마이페이지 */}
       <Route path="/mypage" element={<MyPage />}>
         <Route path="my-info" element={<MyInfo />} />
         <Route path="my-meetings" element={<MyMeetings />} />
         <Route path="change-password" element={<ChangePassword />} />
         <Route path="account-deletion" element={<AccountDeletion />} />
       </Route>
+      {/* 주최한 모임 -> 신청자 보기 페이지 */}
       <Route path="/view-applicant/:id" element={<ViewParticipantPage />} />
+      {/* 주최한 모임 -> 신청자 보기 페이지 -> 신청자 프로필 보기 */}
+      <Route path="/view-applicant/profile/:id" element={<UserProfilePage />} />
+      {/* 채팅 유저 프로필 보기 */}
+      <Route path="/chat/profile/:id" element={<UserProfilePage />} />
     </Routes>
   );
 };
